% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proxfa.R
\name{proxfa}
\alias{proxfa}
\title{Proximal gradient descent for factor analysis}
\usage{
proxfa(Y, L, D, mu, tau, step_size, step_size_shrink = 0.5,
  max_iter = 100, max_inner_iter = 20, tol = 0.001,
  inner_tol = 0.001, n_print = 10, prox_fn = function(L) {     L },
  eps = 1e-08, fix_D = FALSE, fix_mu = FALSE, fix_tau = FALSE)
}
\arguments{
\item{Y}{is the n x p data matrix with no missing entries}

\item{L}{is the intial n x Q loadings matrix}

\item{D}{is the intial Q x Q diagonal matrix of prior variances}

\item{mu}{is the intial p x 1 vector of feature specific means}

\item{tau}{is the intial the residual precision}

\item{step_size}{is the intial step_size for proximal 
gradient descent with line search}

\item{step_size_shrink}{is a value between 0 and 1 to 
shrink the step-size in each line search inner iteration}

\item{max_iter}{is the maximum number of iterations to run
the algorithim}

\item{max_inner_iter}{is the maximum number of inner iterations for
the inner loop of proximal gradient descent on the loadings}

\item{tol}{is the convergence tolerance for the change in the loss}

\item{inner_tol}{is the convergence tolerance for the change in loss for 
the inner loop of proximal gradient descent on the loadings}

\item{n_print}{is the interval of iterations before printing an update}

\item{prox_fn}{is a function that implements the proximal operation on L
by default is is the identity}

\item{eps}{is a small number for dealing with numerical precision}

\item{fix_D}{is a boolean to update the prior variances}

\item{fix_mu}{is a boolean to update the means}

\item{fix_tau}{is a boolean to update the residual precision}
}
\value{
returns a list with fitted values
}
\description{
IN PROGRESS
}
